requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
chromadb>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.10
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
